requirements.txt
fastapi
uvicorn[standard]
pydantic
boto3
python-dotenv
sqlalchemy
pymysql
cryptography
requests
